Move create squad request tests to service tests
