+++ spec/domain/formations_spec.exs
@@ -57,74 +57,39 @@ defmodule Squadster.Domain.FormationsSpec do
   describe "create_squad_request/2" do
     let :squad, do: insert(:squad)
 
-    it "creates a new squad_request" do
+    it "returns new squad_request" do
+      {:ok, squad_request} = Formations.create_squad_request(squad().id, user())
+      expect(squad_request.__struct__) |> to(eq Squadster.Formations.SquadRequest)
+    end
+  end
+
+  describe "delete_squad_request/2" do
+    let! :squad_request, do: insert(:squad_request, user: user())
+
+    it "deletes existing squad_request" do
       previous_count = entities_count(SquadRequest)
-      Formations.create_squad_request(squad().id, user())
-      expect entities_count(SquadRequest) |> to(eq previous_count + 1)
+      Formations.delete_squad_request(squad_request().id, user())
+      expect entities_count(SquadRequest) |> to(eq previous_count - 1)
     end
+  end
 
-    context "when user has another request" do
-      it "should delete old request and create new one" do
-        count = entities_count(SquadRequest)
-        {:ok, %{id: id}} = Formations.create_squad_request(squad().id, user())
+  describe "approve_squad_request/2" do
+    let! :squad_request, do: insert(:squad_request, user: insert(:user), squad: squad())
+    let :squad, do: build(:squad) |> with_commander(user()) |> insert
 
-        expect entities_count(SquadRequest) |> to(eq count + 1)
-
-        count = entities_count(SquadRequest)
-        {:ok, %{id: new_id}} = Formations.create_squad_request(squad().id, user())
-
-        expect entities_count(SquadRequest) |> to(eq count)
-        expect new_id |> to_not(eq id)
+    context "when user has enough permissions" do
+      it "returns new squad_member" do
+        {:ok, squad_member} = Formations.approve_squad_request(squad_request().id, user())
+        expect(squad_member.__struct__) |> to(eq Squadster.Formations.SquadMember)
       end
     end
 
-    context "when user has a squad" do
-      before do
-        insert(:squad_member, user: user(), squad: squad())
-      end
+    context "when user does not have enough permissions" do
+      let :squad, do: insert(:squad)
 
-      it "should not create request" do
-        initial_count = entities_count(SquadRequest)
-
-        Formations.create_squad_request(squad().id, user())
-
-        expect entities_count(SquadRequest) |> to(eq initial_count)
-      end
-
-      it "should return error with message" do
-        {:error, message} = Formations.create_squad_request(squad().id, user())
-        expect message |> to_not(be_nil())
-      end
-    end
-
-    describe "delete_squad_request/2" do
-      let! :squad_request, do: insert(:squad_request, user: user())
-
-      it "deletes existing squad_request" do
-        previous_count = entities_count(SquadRequest)
-        Formations.delete_squad_request(squad_request().id, user())
-        expect entities_count(SquadRequest) |> to(eq previous_count - 1)
-      end
-    end
-
-    describe "approve_squad_request/2" do
-      let! :squad_request, do: insert(:squad_request, user: insert(:user), squad: squad())
-      let :squad, do: build(:squad) |> with_commander(user()) |> insert
-
-      context "when user has enough permissions" do
-        it "returns new squad_member" do
-          {:ok, squad_member} = Formations.approve_squad_request(squad_request().id, user())
-          expect(squad_member.__struct__) |> to(eq Squadster.Formations.SquadMember)
-        end
-      end
-
-      context "when user does not have enough permissions" do
-        let :squad, do: insert(:squad)
-
-        it "returns error" do
-          {:error, message} = Formations.approve_squad_request(squad_request().id, user())
-          expect message |> to_not(be nil)
-        end
+      it "returns error" do
+        {:error, message} = Formations.approve_squad_request(squad_request().id, user())
+        expect message |> to_not(be nil)
       end
     end
   end
